Modified delete button controller in builder to prevent errors when there are no more levels.

+++ src/main/java/SixesWild/com/mimas/sixeswild/controllers/BuilderDeleteUserLevelButtonController.java
@@ -37,24 +37,29 @@ public class BuilderDeleteUserLevelButtonController implements ActionListener {
 	 * java.awt.event.ActionListener#actionPerformed(java.awt.event.ActionEvent)
 	 */
 	public void actionPerformed(ActionEvent arg0) {
-		String fileName = ((String) this.app.getBuilderView()
-				.getBuilderTopPanel().getDeleteComboBox().getSelectedItem())
-				.substring(3);
-		fileName += ".xml";
-		fileName = fileName.substring(fileName.indexOf(" ") + 1);
-		logger.log(Level.INFO, fileName);
 
-		// Open the file to delete
-		try {
-			File level = new File(XMLParser.USER_DIR + "/" + fileName);
+		// Check if there are even files left to delete
+		if (this.app.getBuilderView().getBuilderTopPanel().getDeleteComboBox()
+				.getItemCount() > 0) {
+			String fileName = ((String) this.app.getBuilderView()
+					.getBuilderTopPanel().getDeleteComboBox().getSelectedItem())
+					.substring(3);
+			fileName += ".xml";
+			fileName = fileName.substring(fileName.indexOf(" ") + 1);
+			logger.log(Level.INFO, fileName);
 
-			if (level.delete()) {
-				logger.log(Level.INFO, "Deleted successfully.");
-			} else {
-				logger.log(Level.INFO, "Delete failed.");
+			// Open the file to delete
+			try {
+				File level = new File(XMLParser.USER_DIR + "/" + fileName);
+
+				if (level.delete()) {
+					logger.log(Level.INFO, "Deleted successfully.");
+				} else {
+					logger.log(Level.INFO, "Delete failed.");
+				}
+			} catch (Exception e) {
+				logger.log(Level.WARNING, "File not found.");
 			}
-		} catch (Exception e) {
-			logger.log(Level.WARNING, "File not found.");
 		}
 
 		// Update the level lists
